improve config api and cli
